config/requirements.txt
Django==1.8.5
Pillow==3.0.0
argparse==1.2.1
djangorestframework==3.3.1
wsgiref==0.1.2
